Started implementing tests
