+++ dataland-frontend/tests/e2e/specs/company-metadata/SearchCompaniesForFrameworkData.ts
@@ -73,25 +73,25 @@ describe("As a user, I expect the search functionality on the /companies page to
   it("Scroll the page and check if search icon and search bar behave as expected", () => {
     cy.visitAndCheckAppMount("/companies");
     cy.get("input[name=search_bar_top]").type("a").type("{enter}");
-    cy.get("button[name=search_bar_collapse]").should("not.exist");
+    cy.get("button[name=search_bar_collapse]").should("not.be.visible");
     verifyPaginator();
 
     cy.scrollTo(0, 500, { duration: 300 });
     cy.get("input[name=search_bar_top]").should("exist");
-    cy.get("button[name=search_bar_collapse]").should("exist");
+    cy.get("button[name=search_bar_collapse]").should("be.visible");
 
     cy.scrollTo(0, 0, { duration: 300 });
     cy.get("input[name=search_bar_top]").should("exist");
-    cy.get("button[name=search_bar_collapse]").should("not.exist");
+    cy.get("button[name=search_bar_collapse]").should("not.be.visible");
 
     cy.scrollTo(0, 500, { duration: 300 });
     cy.get("button[name=search_bar_collapse]").should("exist").click();
     cy.get("input[name=search_bar_top]").should("not.exist");
     cy.get("input[name=search_bar_scrolled]").should("exist");
-    cy.get("button[name=search_bar_collapse]").should("not.exist");
+    cy.get("button[name=search_bar_collapse]").should("not.be.visible");
 
     cy.scrollTo(0, 480, { duration: 300 });
-    cy.get("button[name=search_bar_collapse]").should("exist");
+    cy.get("button[name=search_bar_collapse]").should("be.visible");
     cy.get("input[name=search_bar_top]").should("exist");
     cy.get("input[name=search_bar_scrolled]").should("not.exist");
   });
@@ -187,93 +187,3 @@ describe("As a user, I expect the search functionality on the /companies page to
     });
   });
 });
-
-describeIf(
-  "As a user, I expect the search functionality on the /companies page to adjust to the framework filters",
-  {
-    executionEnvironments: ["developmentLocal", "development"],
-    dataEnvironments: ["fakeFixtures"],
-  },
-  function () {
-    beforeEach(function () {
-      cy.ensureLoggedIn("data_uploader", Cypress.env("KEYCLOAK_UPLOADER_PASSWORD"));
-    });
-
-    it(
-      "Upload a company without uploading framework data for it and check if it neither appears in the + " +
-        "autocomplete suggestions nor in the search results, even though no framework filter is set.",
-      () => {
-        const companyName = "ThisCompanyShouldNeverBeFound12349876";
-        createCompanyAndGetId(companyName);
-        cy.visit(`/companies`);
-        cy.intercept("**/api/companies*").as("searchCompany");
-        cy.get("input[name=search_bar_top]").click({ force: true }).type(companyName);
-        cy.wait("@searchCompany", { timeout: 2 * 1000 }).then(() => {
-          cy.get(".p-autocomplete-item").should("not.exist");
-        });
-        cy.visit(`/companies?input=${companyName}`)
-          .get("div[class='col-12 text-left']")
-          .should("contain.text", "Sorry! Your search didn't return any results.");
-      }
-    );
-
-    const companyNameMarker = "Data987654321";
-
-    it(
-      "Upload a company with Eu Taxonomy Data For Financials and check if it only appears in the results if the " +
-        "framework filter is set to that framework, or to several frameworks including that framework",
-      () => {
-        const companyName = "CompanyWithFinancial" + companyNameMarker;
-        createCompanyAndGetId(companyName).then((companyId) => uploadDummyEuTaxonomyDataForFinancials(companyId));
-        cy.visit(`/companies?input=${companyName}`)
-          .get("td[class='d-bg-white w-3 d-datatable-column-left']")
-          .contains(companyName)
-          .should("exist");
-        cy.visit(`/companies?input=${companyName}&frameworks=eutaxonomy-financials`)
-          .get("td[class='d-bg-white w-3 d-datatable-column-left']")
-          .contains(companyName)
-          .should("exist");
-        cy.visit(`/companies?input=${companyName}&framework=eutaxonomy-non-financials`)
-          .get("div[class='col-12 text-left']")
-          .should("contain.text", "Sorry! Your search didn't return any results.");
-        cy.visit(`/companies?input=${companyName}&framework=eutaxonomy-non-financials&framework=eutaxonomy-financials`)
-          .get("td[class='d-bg-white w-3 d-datatable-column-left']")
-          .contains(companyName)
-          .should("exist");
-      }
-    );
-
-    function checkFirstAutoCompleteSuggestion(companyNamePrefix: string, frameworkToFilterFor: string): void {
-      cy.visit(`/companies?frameworks=${frameworkToFilterFor}`);
-      cy.intercept("**/api/companies*").as("searchCompany");
-      cy.get("input[name=search_bar_top]").click({ force: true }).type(companyNameMarker);
-      cy.wait("@searchCompany", { timeout: 2 * 1000 }).then(() => {
-        cy.get(".p-autocomplete-item")
-          .eq(0)
-          .get("span[class='font-normal']")
-          .contains(companyNamePrefix)
-          .should("exist");
-      });
-    }
-
-    it(
-      "Upload a company with Eu Taxonomy Data For Financials and one with Eu Taxonomy Data For Non-Financials and " +
-        "check if they are displayed in the autocomplete dropdown only if the framework filter is set accordingly",
-      () => {
-        const companyNameFinancialPrefix = "CompanyWithFinancial";
-        const companyNameFinancial = companyNameFinancialPrefix + companyNameMarker;
-        createCompanyAndGetId(companyNameFinancial).then((companyId) =>
-          uploadDummyEuTaxonomyDataForFinancials(companyId)
-        );
-        checkFirstAutoCompleteSuggestion(companyNameFinancialPrefix, "eutaxonomy-financials");
-
-        const companyNameNonFinancialPrefix = "CompanyWithNonFinancial";
-        const companyNameNonFinancial = companyNameNonFinancialPrefix + companyNameMarker;
-        createCompanyAndGetId(companyNameNonFinancial).then((companyId) =>
-          uploadDummyEuTaxonomyDataForNonFinancials(companyId)
-        );
-        checkFirstAutoCompleteSuggestion(companyNameNonFinancialPrefix, "eutaxonomy-non-financials");
-      }
-    );
-  }
-);
